+++ platypush/plugins/music/__init__.py
@@ -6,45 +6,80 @@ class MusicPlugin(Plugin):
         super().__init__(**kwargs)
 
     @action
-    def play(self):
+    def play(self, **kwargs):
         raise NotImplementedError()
 
     @action
-    def pause(self):
+    def pause(self, **kwargs):
         raise NotImplementedError()
 
     @action
-    def stop(self):
+    def stop(self, **kwargs):
         raise NotImplementedError()
 
     @action
-    def next(self):
+    def next(self, **kwargs):
         raise NotImplementedError()
 
     @action
-    def previous(self):
+    def previous(self, **kwargs):
         raise NotImplementedError()
 
     @action
-    def set_volume(self, volume):
+    def set_volume(self, volume, **kwargs):
         raise NotImplementedError()
 
     @action
-    def seek(self, position):
+    def volup(self, delta, **kwargs):
         raise NotImplementedError()
 
     @action
-    def add(self, content):
+    def voldown(self, delta, **kwargs):
         raise NotImplementedError()
 
     @action
-    def clear(self):
+    def seek(self, position, **kwargs):
         raise NotImplementedError()
 
     @action
-    def status(self):
+    def add(self, resource, **kwargs):
+        raise NotImplementedError()
+
+    @action
+    def clear(self, **kwargs):
+        raise NotImplementedError()
+
+    @action
+    def status(self, **kwargs):
+        raise NotImplementedError()
+
+    @action
+    def current_track(self, **kwargs):
+        raise NotImplementedError()
+
+    @action
+    def get_playlists(self, **kwargs):
+        raise NotImplementedError()
+
+    @action
+    def get_playlist(self, playlist, **kwargs):
+        raise NotImplementedError()
+
+    @action
+    def add_to_playlist(self, playlist, resources, **kwargs):
+        raise NotImplementedError()
+
+    @action
+    def remove_from_playlist(self, playlist, resources, **kwargs):
+        raise NotImplementedError()
+
+    @action
+    def playlist_move(self, playlist, from_pos: int, to_pos: int, **kwargs):
+        raise NotImplementedError()
+
+    @action
+    def search(self, query, *args, **kwargs):
         raise NotImplementedError()
 
 
 # vim:sw=4:ts=4:et:
-